requirements.txt
langchain
langgraph
langchain-community
langchain-groq
pinecone-client
pypdf
sentence-transformers
streamlit